# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tree State Management API",
    "version": "1.0.0",
    "description": "REST API for managing hierarchical tree data structures"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Nodes",
      "description": "Operations for managing nodes in the tree"
    }
  ],
  "paths": {
    "/nodes": {
      "get": {
        "summary": "Get all nodes",
        "operationId": "getNodes",
        "tags": ["Nodes"],
        "responses": {
          "200": {
            "description": "Successfully retrieved all nodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TreeNode"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new node",
        "operationId": "addNode",
        "tags": ["Nodes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["node", "parentNodeId"],
                "properties": {
                  "node": {
                    "$ref": "#/components/schemas/TreeNodeProperties"
                  },
                  "parentNodeId": {
                    "type": "string"
                  },
                  "insertAtIndex": {
                    "type": "integer",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Node successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TreeNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/nodes/{nodeId}": {
      "patch": {
        "summary": "Update a node's properties",
        "operationId": "updateNode",
        "tags": ["Nodes"],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeNodeProperties"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TreeNode"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a node",
        "operationId": "removeNode",
        "tags": ["Nodes"],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Node successfully removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TreeNode"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          }
        }
      }
    },
    "/nodes/{nodeId}/parent": {
      "put": {
        "summary": "Change a node's parent",
        "operationId": "setNodeParent",
        "tags": ["Nodes"],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newParentId"],
                "properties": {
                  "newParentId": {
                    "type": "string"
                  },
                  "insertAtIndex": {
                    "type": "integer",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node's parent successfully changed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TreeNode"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          },
          "400": {
            "description": "Invalid parent (e.g. would create circular reference)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TreeNodeProperties": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the node"
          },
          "description": {
            "type": "string",
            "description": "A detailed description of the node"
          },
          "readinessLevel": {
            "type": "integer",
            "description": "User-set readiness level (1-9)",
            "minimum": 1,
            "maximum": 9,
            "nullable": true
          }
        }
      },
      "TreeNode": {
        "type": "object",
        "required": ["id", "childrenIds", "parentId"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the node"
          },
          "parentId": {
            "type": "string",
            "description": "ID of the parent node. Null for root node.",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "The title of the node"
          },
          "description": {
            "type": "string",
            "description": "A detailed description of the node"
          },
          "readinessLevel": {
            "type": "integer",
            "description": "User-set readiness level (1-9)",
            "minimum": 1,
            "maximum": 9,
            "nullable": true
          },
          "readinessLevelCalculated": {
            "type": "integer",
            "description": "Calculated readiness level based on children (1-9)",
            "minimum": 1,
            "maximum": 9
          },
          "childrenIds": {
            "type": "array",
            "description": "Array of child node IDs",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}